refactor(pace): retire the unused Codex v2 evaluator
Remove the private agent_history module, including its dormant schema-2 writer, persistence path, historical evaluator, and evaluator-only tests. The module had no production caller; provider quota cards already use the schema-3 implementation in agent_quota_history.

Move the deterministic weighted_median helper into agent_quota_history with its existing mismatch, zero-weight, ordering, and exact-half semantics intact. The three production fit/projection callers and the existing tie regression now use the private local helper.

Preserve the current-account-bound v2-to-v3 importer, its LegacyV2Store and LegacyV2Sample parser, collision and idempotence rules, legacy reader, storage behavior, and the successful Codex request caller without semantic changes. V1 remains untouched and codex-weekly-history-v2.json remains a read-only migration input.

Update the superseded v2 plan, active provider pace plan, and verification contract so the retired evaluator is historical evidence rather than a current gate. The shipping schema-3 evaluator and importer regressions remain the authoritative runtime checks.

Verification: focused agent_quota_history tests passed (77 passed, 1 ignored); full Rust workspace tests passed; Clippy, make build, make selftest, knowledge checks, and diff checks passed. Workspace rustfmt reported the same nine pre-existing baseline paths. The replayed patch matched candidate fingerprint 155fee7307e35751cbf57e29bd94f443ea0794d2630941ec806ba8400b100610, and fresh outcome verification returned CONFIRMED.
